Account for tabs in ErrorContext
Prior to this, there was a bug where tabs would only be counted as one
column, and therefore the error indicating carrot (^) would point to a
much earlier column if the line included tabs.

This is solved by accounting for tabs in the column count, and also
converting any tabs in the line into spaces when they get output. At
first this was perceived as an unnecessary allocation, but then I
realized that at scale, each ErrorContext should allocate its own line,
since using the input buffer would require that the input buffer exist
until all ErrorContexts are printed. And since the input buffer is the
file, that would mean that all input files' contents would need to
remain in memory until the ErrorContexts get printed. This is obviously
not scalable, and so ErrorContext owning its line solves both problems.